refactor: 이동관련 MoveStrategy로 추상화
