+++ src/main/java/chess/Strategy/KingMoveStrategy.java
@@ -1,0 +1,46 @@
+package chess.Strategy;
+
+import chess.domain.board.Board;
+import chess.domain.board.Horizontal;
+import chess.domain.board.Position;
+import chess.domain.board.Vertical;
+import chess.domain.piece.Direction;
+import chess.domain.piece.Piece;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+public class KingMoveStrategy extends SpecifiedMoveStrategy {
+    @Override
+    public List<Position> searchMovablePositions(Position target) {
+        List<Position> movablePositions = new ArrayList<>();
+
+        int horizontalWeight = target.getHorizontalWeight();
+        int verticalWeight = target.getVerticalWeight();
+
+        for (Direction direction : Direction.values()) {
+            int changedHorizontalWeight = horizontalWeight + direction.getX();
+            int changedVerticalWeight = verticalWeight + direction.getY();
+            movablePositions.addAll(calculateBoardPosition(changedHorizontalWeight, changedVerticalWeight));
+        }
+
+        return movablePositions;
+    }
+
+    private List<Position> calculateBoardPosition(int changedHorizontalWeight, int changedVerticalWeight) {
+        List<Position> result = new ArrayList<>();
+        if (isInBorder(changedHorizontalWeight, changedVerticalWeight)) {
+            result.add(Position.of(
+                    Horizontal.findFromWeight(changedHorizontalWeight),
+                    Vertical.findFromWeight(changedVerticalWeight))
+            );
+        }
+        return result;
+    }
+
+    private boolean isInBorder(int horizontalWeight, int verticalWeight) {
+        return horizontalWeight >= Board.MIN_BORDER && horizontalWeight <= Board.MAX_BORDER
+                && verticalWeight >= Board.MIN_BORDER && verticalWeight <= Board.MAX_BORDER;
+    }
+}
