+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,65 +1,12 @@
 package chess.domain.piece;
 
-import chess.domain.board.Board;
-import chess.domain.board.Horizontal;
-import chess.domain.board.Position;
-import chess.domain.board.Vertical;
-
-import java.util.ArrayList;
-import java.util.List;
+import chess.Strategy.BishopMoveStrategy;
 
 public class Bishop extends Piece {
     private static final String BISHOP_NAME = "B";
     private static final double SCORE = 3;
 
     public Bishop(Team team) {
-        super(BISHOP_NAME, team, SCORE);
-    }
-
-    @Override
-    public List<Position> searchMovablePositions(Position target) {
-        List<Position> movablePositions = new ArrayList<>();
-
-        movablePositions.addAll(calculateBishopMovablePositions(target, Direction.RIGHT_TOP));
-        movablePositions.addAll(calculateBishopMovablePositions(target, Direction.LEFT_TOP));
-        movablePositions.addAll(calculateBishopMovablePositions(target, Direction.LEFT_BOTTOM));
-        movablePositions.addAll(calculateBishopMovablePositions(target, Direction.RIGHT_BOTTOM));
-
-        return movablePositions;
-    }
-
-    private List<Position> calculateBishopMovablePositions(Position target, Direction direction) {
-        List<Position> result = new ArrayList<>();
-        int horizontalWeight = target.getHorizontalWeight();
-        int verticalWeight = target.getVerticalWeight();
-        while (horizontalWeight + direction.getX() >= Board.MIN_BORDER && horizontalWeight + direction.getX() <= Board.MAX_BORDER
-                && verticalWeight + direction.getY() >= Board.MIN_BORDER && verticalWeight + direction.getY() <= Board.MAX_BORDER) {
-            horizontalWeight += direction.getX();
-            verticalWeight += direction.getY();
-            result.add(
-                    Position.of(Horizontal.findFromWeight(horizontalWeight), Vertical.findFromWeight(verticalWeight))
-            );
-        }
-        return result;
-    }
-
-    @Override
-    public boolean canMove(Position target, Position destination, Board board) {
-        Direction direction = target.directionToDestination(destination);
-        Position movedPosition = target;
-        while (true) {
-            movedPosition = movedPosition.moveTowardDirection(direction);
-
-            if (movedPosition != destination) {
-                if (board.findPieceFromPosition(movedPosition) != null) {
-                    return false;
-                }
-            }
-            if (movedPosition == destination) {
-                Piece targetPiece = board.findPieceFromPosition(target);
-                Piece destinationPiece = board.findPieceFromPosition(movedPosition);
-                return destinationPiece == null || !destinationPiece.isSameTeam(targetPiece);
-            }
-        }
+        super(BISHOP_NAME, team, SCORE, new BishopMoveStrategy());
     }
 }
