+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,5 +1,6 @@
 package chess.domain.piece;
 
+import chess.Strategy.MoveStrategy;
 import chess.domain.board.Board;
 import chess.domain.board.Position;
 
@@ -9,11 +10,13 @@ public abstract class Piece {
     private final String name;
     private final Team team;
     private final double score;
+    private final MoveStrategy moveStrategy;
 
-    public Piece(String name, Team team, double score) {
+    public Piece(String name, Team team, double score, MoveStrategy moveStrategy) {
         this.team = team;
         this.name = convertName(name, team);
         this.score = score;
+        this.moveStrategy = moveStrategy;
     }
 
     private String convertName(String name, Team team) {
@@ -23,21 +26,13 @@ public abstract class Piece {
         return name.toLowerCase();
     }
 
-    public String getName() {
-        return name;
+    public boolean canMove(Position target, Position destination, Board board) {
+        return moveStrategy.canMove(target, destination, board);
     }
 
-    public boolean isBlack() {
-        return team.isBlack();
+    public List<Position> searchMovablePositions(Position target) {
+        return moveStrategy.searchMovablePositions(target);
     }
-
-    public boolean isWhite() {
-        return team.isWhite();
-    }
-
-    public abstract boolean canMove(Position target, Position destination, Board board);
-
-    public abstract List<Position> searchMovablePositions(Position target);
 
     public boolean isSameTeam(Piece piece) {
         return team.equals(piece.getTeam());
@@ -45,6 +40,10 @@ public abstract class Piece {
 
     public boolean isSameTeam(Team team) {
         return this.team.equals(team);
+    }
+
+    public String getName() {
+        return name;
     }
 
     public Team getTeam() {
